Working on User.js
