+++ my-types.d.ts
@@ -10,3 +10,24 @@ export type ErrorRequestHandler = (
 	res: Response,
 	next: NextFunction
 ) => void;
+
+//db.ts
+interface ssl {
+	rejectUnauthorized: boolean;
+}
+interface DialectOptions {
+	ssl: ssl;
+}
+
+export interface Config {
+	logging?: boolean;
+	dialectOptions?: DialectOptions;
+}
+
+// export interface ResponseError extends Error {
+// 	status?: number;
+// }
+
+export interface Error {
+	status?: number;
+}
